refactor: 리팩터링 요구 사항에 따라 Car 테스트 코드에서 jest.fn을 사용하지 않도록 수정
