+++ __tests__/domain/Car.test.js
@@ -1,19 +1,11 @@
-import Car from "../../src/models/Car";
-import gameUtils from "../../src/utils/gameUtils";
+import Car from '../../src/domain/Car';
 
-const mockRandoms = (numbers) => {
-  gameUtils.pickRandomNumber = jest.fn();
-  numbers.reduce((acc, number) => {
-    return acc.mockReturnValueOnce(number);
-  }, gameUtils.pickRandomNumber);
-}
-
-describe('자동차 테스트', () => {
+describe('자동차 도메인 테스트', () => {
   describe('자동차 생성 테스트', () => {
     test('자동차 이름이 비었으면 에러를 발생시킨다.', () => {
       // given
       const invalidInput = '';
-      
+
       expect(() => {
         // when
         new Car(invalidInput);
@@ -23,7 +15,7 @@ describe('자동차 테스트', () => {
     test('자동차 이름이 5자 초과되면 에러를 발생시킨다.', () => {
       // given
       const invalidInput = 'cookie';
-      
+
       expect(() => {
         // when
         new Car(invalidInput);
@@ -33,7 +25,7 @@ describe('자동차 테스트', () => {
     test('자동차 이름이 한글 또는 영어가 아니면 에러를 발생시킨다.', () => {
       // given
       const invalidInput = 'pobi1';
-      
+
       expect(() => {
         // when
         new Car(invalidInput);
@@ -51,32 +43,30 @@ describe('자동차 테스트', () => {
       expect(typeof car === 'object').toBeTruthy();
     });
   });
-  describe('자동차 전진테스트, (4이상이면 전진)', () => {
+  describe('자동차 전진테스트, (전진 조건: 랜덤한 수 4 이상, 전진: 1, 멈춤: 0)', () => {
     const sundayCar = new Car('썬데이');
 
-    test('랜덤한 수를 3을 받으면 전진하지 않습니다.', () => {
+    test('랜덤한 수 3을 받으면 전진하지 않는다.', () => {
       // given
-      const randomNumber = [3];
+      const randomNumber = 3;
       const answer = 0;
-      mockRandoms(randomNumber);
 
       // when
-      const actResult = sundayCar.actCar();
+      const setResult = sundayCar.setCar(randomNumber);
 
       // then
-      expect(actResult.score).toEqual(answer);
+      expect(setResult.score).toEqual(answer);
     });
-    test('랜덤한 수를 4을 받으면 전진합니다.', () => {
+    test('랜덤한 수 4를 받으면 전진한다.', () => {
       // given
-      const randomNumber = [4];
+      const randomNumber = 4;
       const answer = 1;
-      mockRandoms(randomNumber);
 
       // when
-      const actResult = sundayCar.actCar();
+      const setResult = sundayCar.setCar(randomNumber);
 
       // then
-      expect(actResult.score).toEqual(answer);
+      expect(setResult.score).toEqual(answer);
     });
   });
 });
